+++ config/oauth2ra.go
@@ -5,6 +5,12 @@ import (
 	"github.com/hashicorp/hcl/v2/gohcl"
 )
 
+const (
+	ClientCredentials = "client_credentials"
+	JwtBearer         = "urn:ietf:params:oauth:grant-type:jwt-bearer"
+	Password          = "password"
+)
+
 var OAuthBlockSchema = &hcl.BodySchema{
 	Blocks: []hcl.BlockHeaderSchema{
 		{
@@ -22,17 +28,18 @@ var (
 
 // OAuth2ReqAuth represents the oauth2 block in a backend block.
 type OAuth2ReqAuth struct {
-	BackendName             string   `hcl:"backend,optional"`
-	ClientID                string   `hcl:"client_id"`
-	ClientSecret            string   `hcl:"client_secret"`
-	GrantType               string   `hcl:"grant_type"`
-	Password                string   `hcl:"password,optional"` // password undocumented feature!
-	Remain                  hcl.Body `hcl:",remain"`
-	Retries                 *uint8   `hcl:"retries,optional"`
-	Scope                   string   `hcl:"scope,optional"`
-	TokenEndpoint           string   `hcl:"token_endpoint,optional"`
-	TokenEndpointAuthMethod *string  `hcl:"token_endpoint_auth_method,optional"`
-	Username                string   `hcl:"username,optional"` // username undocumented feature!
+	AssertionExpr           hcl.Expression `hcl:"assertion,optional"`
+	BackendName             string         `hcl:"backend,optional"`
+	ClientID                string         `hcl:"client_id,optional"`
+	ClientSecret            string         `hcl:"client_secret,optional"`
+	GrantType               string         `hcl:"grant_type"`
+	Password                string         `hcl:"password,optional"`
+	Remain                  hcl.Body       `hcl:",remain"`
+	Retries                 *uint8         `hcl:"retries,optional"`
+	Scope                   string         `hcl:"scope,optional"`
+	TokenEndpoint           string         `hcl:"token_endpoint,optional"`
+	TokenEndpointAuthMethod *string        `hcl:"token_endpoint_auth_method,optional"`
+	Username                string         `hcl:"username,optional"`
 }
 
 // Reference implements the <BackendReference> interface.
@@ -69,6 +76,10 @@ func (oa *OAuth2ReqAuth) Schema(inline bool) *hcl.BodySchema {
 	}
 
 	return schema
+}
+
+func (oa *OAuth2ReqAuth) ClientAuthenticationRequired() bool {
+	return oa.GrantType != JwtBearer
 }
 
 func (oa *OAuth2ReqAuth) GetClientID() string {